Split gif/mp4 function

+++ vidcutter/libs/videoservice.py
@@ -224,7 +224,7 @@ class VideoService(QObject):
         acodec = re.search(r'Stream.*Audio:\s(\w+)', result).group(1)
         return vcodec, acodec
 
-    def cut(self, source: str, output: str, frametime: str, duration: str, allstreams: bool = True, gifonly: bool = False) -> bool:
+    def cut(self, source: str, output: str, frametime: str, duration: str, allstreams: bool = True, gifOutput: bool = False, mp4Output: bool = False) -> bool:
         self.checkDiskSpace(output)
         args = '-y -ss {0} -i "{1}" -t {2} -c:v libx264 -an -crf 22 -pix_fmt yuv420p ' + \
             '-vf "scale=iw*min(1\,min(1280/iw\,720/ih)):-2" -preset fast "{3}"'
@@ -238,14 +238,15 @@ class VideoService(QObject):
         cut_arg = '-ss {0} -i "{1}" -t {2} -vf "fps=24,scale=iw*min(1\,min(640/iw\,360/ih)):-2" "{3}"'
         merged_arg = '{0} --fps 24 --output "{1}"'
 
-        os.mkdir(png_dir)
-        self.cmdExec(self.backend,
-                     cut_arg.format(frametime, source, duration, png_dir + png_frame))
-        self.cmdExec(self.gifski,
-                     merged_arg.format(png_dir + '*.png', gif_dir))
-        shutil.rmtree(png_dir)
+        if gifOutput:
+            os.mkdir(png_dir)
+            self.cmdExec(self.backend,
+                         cut_arg.format(frametime, source, duration, png_dir + png_frame))
+            self.cmdExec(self.gifski,
+                         merged_arg.format(png_dir + '*.png', gif_dir))
+            shutil.rmtree(png_dir)
 
-        if not gifonly:
+        if mp4Output:
             self.cmdExec(self.backend, args.format(frametime, source, duration, QDir.fromNativeSeparators(output)))
 
         return True
